Added interface FileReaderService
